found sources
Found a couple sources for generating regular n-simplices
