+++ convexHullOperations.py
@@ -1,0 +1,7 @@
+
+
+
+def selectDirections(dims):
+	# https://math.stackexchange.com/questions/2533532/how-to-find-the-coordinates-of-an-n-dimensional-rectangular-simplex
+	# https://en.wikipedia.org/wiki/Simplex
+	pass
