+++ hullMain.py
@@ -1,47 +1,11 @@
-from typing_extensions import Concatenate
-from matrixOperations import *
+from barycentric import *
 
 
-def hullMainScript(points):
-	pass
-
-
-def pointInside(simplex, point, simplexMat):
-	lambdaVec = barycentric(simplex, point, simplexMat)
-	isInside = True
-	for lam in lambdaVec[0]:
-		if lam < 0: isInside = False
-	return isInside
-
-
-def simplexMatrix(simplex):
-	dimensions = len(simplex) - 1
-	matT = [[simplex[j][i] - simplex[-1][i] for j in range(len(simplex) - 1)] for i in range(dimensions)]
-	simpMat = matInverse(matT)
-	return simpMat
-
-
-def barycentric(simplex, point, simplexMat):
-	# source: https://en.wikipedia.org/wiki/Barycentric_coordinate_system 
-	# section: Barycentric coordinates on Tetrahedra
-	dimensions = len(point[0])
-	multVec = transpose([[point[0][i] - simplex[-1][i] for i in range(dimensions)]])
-	lambdaVec = matMul(simplexMat, multVec)
-	finalLambda = 1 - sum(lambdaVec[0])
-	lambdaVec[0].append(finalLambda)
-	return lambdaVec 
-
-
-
-def calcDistance(vector, point):
-	distance = sum(point[0][comp]*vector[0][comp] for comp in range(len(point[0])))
-	return distance
-
-
-def unitVec(vector):
-	length = pow(sum(comp*comp for comp in vector[0]), 1/2)
-	unitVec = [comp/length for comp in vector[0]]
-	return unitVec
+def hullMainScript():
+	simplex = [[0, 0, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0], [0, 1, 0, 0], [1, 0, 0, 0]]
+	simpMat = simplexMatrix(simplex)
+	point = [[0, 0.5001, 0.5, 0]]
+	print(pointInside(simplex, point, simpMat))
 
 
 if __name__ == '__main__':
